refactor(ovh-location-tile): update name for consistency and new docs
