+++ package.json
@@ -37,7 +37,8 @@
     "doc": "lerna run doc",
     "doc:api": "lerna run doc:api",
     "doc:api:slow": "lerna run --concurrency 1 doc:api",
-    "generate:component": "yarn workspace @ovhcloud/ods-component-generator run generate",
+    "generate:component": "plop --plopfile scripts/component-generator/plopfile.js component",
+    "generate:component0000": "yarn workspace @ovhcloud/ods-component-generator run generate",
     "generate:licence": "npx generate-license-file --input package.json --output THIRD-PARTY-LICENCES --overwrite && lerna run generate:licence",
     "generate:theme": "lerna run generate:theme",
     "ignore:rm": "lerna run ignore:rm",
@@ -105,6 +106,7 @@
     "husky": "4.3.8",
     "lerna": "3.22.1",
     "node-fetch": "2.6.7",
+    "plop": "4.0.0",
     "stylelint": "15.10.3",
     "stylelint-config-concentric-order": "5.2.0",
     "stylelint-config-standard-scss": "11.0.0",
