+++ packages/components-ovh/CHANGELOG.md
@@ -21,7 +21,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 # [16.3.1](https://ovh.github.io/design-system/v16.3.1/?path=/docs/design-system-changelog--page) (2023-11-07)
 [compare](https://github.com/ovh/design-system/compare/v16.3.0...v16.3.1)
 
-**Note:** Version bump only for package @ovhcloud/ods-ovh-components
+**Note:** Version bump only for package @ovhcloud/ods-components-ovh
