+++ packages/components-ovh/location-tile/documentation/specifications/specifications-ovh-location-tile.mdx
