+++ packages/components-ovh/location-tile/documentation/usage-guidelines/usage.mdx
@@ -1,0 +1,12 @@
+import GenericStyle from '@ovhcloud/ods-common-core/docs/generic-style.mdx';
+
+<GenericStyle />
+
+## Usage
+
+### Default
+<osds-ovh-location-tile></osds-ovh-location-tile>
+
+```html
+<osds-ovh-location-tile></osds-ovh-location-tile>
+```
