+++ packages/components-ovh/location-tile/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@ovhcloud/ods-ovh-component-location-tile",
+  "name": "@ovhcloud/ods-component-ovh-location-tile",
   "version": "16.3.2",
-  "description": "ODS OVH Location Tile component",
+  "description": "ODS OvhLocationTile component",
   "private": true,
   "author": "OVH SAS",
   "license": "Apache-2.0",
