+++ packages/components-ovh/location-tile/vue/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@ovhcloud/ods-ovh-component-location-tile-vue",
+  "name": "@ovhcloud/ods-component-ovh-location-tile-vue",
   "version": "16.3.2",
   "private": true,
   "description": "Vue specific wrapper for ods",
@@ -20,7 +20,7 @@
     "dist/"
   ],
   "dependencies": {
-    "@ovhcloud/ods-ovh-component-location-tile": "16.3.2"
+    "@ovhcloud/ods-component-ovh-location-tile": "16.3.2"
   },
   "peerDependencies": {
     "vue": ">=3"
