+++ packages/components-ovh/package.json
@@ -1,12 +1,12 @@
 {
-  "name": "@ovhcloud/ods-ovh-components",
+  "name": "@ovhcloud/ods-components-ovh",
   "version": "16.3.2",
   "description": "ODS OVH Components",
   "private": true,
   "author": "OVH SAS",
   "license": "Apache-2.0",
   "dependencies": {
-    "@ovhcloud/ods-ovh-component-location-tile": "16.3.2"
+    "@ovhcloud/ods-component-ovh-location-tile": "16.3.2"
   },
   "devDependencies": {
     "@ovhcloud/ods-dev": "16.3.2"
