+++ packages/storybook/stories/components-ovh/location-tile/1_specifications.stories.mdx
@@ -1,0 +1,18 @@
+import { Meta } from '@storybook/addon-docs';
+import Specifications from '@ovhcloud/ods-components-ovh/location-tile/documentation/specifications/specifications-ovh-location-tile.mdx';
+
+<Meta title="ODS Components/Content/OvhLocationTile [quark]/Specifications" />
+
+# OvhLocationTile - Design Specifications
+----
+
+OvhLocationTile TODO add small description:
+<div>
+  <osds-ovh-location-tile>
+  </osds-ovh-location-tile>
+</div>
+
+# OvhLocationTile - Technical Specification
+----
+
+<Specifications />
