+++ packages/storybook/stories/components-ovh/location-tile/2_usage.stories.mdx
@@ -1,0 +1,9 @@
+import { Meta } from '@storybook/addon-docs';
+import Usage from '@ovhcloud/ods-components-ovh/location-tile/documentation/usage-guidelines/usage.mdx';
+
+<Meta title="ODS Components/Content/OvhLocationTile [quark]/Usage Guidelines" />
+
+# OvhLocationTile - Usage Guidelines
+----
+
+<Usage />
