--- packages/storybook/stories/components-ovh/location-tile/location-tile.specifications.stories.mdx
@@ -1,9 +1,0 @@
-import { Meta } from '@storybook/addon-docs';
-import SpecificationsLocationTile from '@ovhcloud/ods-ovh-components/location-tile/documentation/specifications/specifications-location-tile.mdx';
-
-<Meta title="OVH Components/Location Tile [molecule]/Specifications" />
-
-# LocationTile - Technical Specification
-----
-
-<SpecificationsLocationTile />
