--- packages/storybook/stories/components-ovh/location-tile/location-tile.web-component.stories.page.mdx
@@ -1,17 +1,0 @@
-import { Description, Meta } from '@storybook/addon-docs';
-import APITable from '@ovhcloud/ods-ovh-components/location-tile/dist/docs/components/osds-ovh-location-tile/readme.md';
-import Usage from '@ovhcloud/ods-ovh-components/location-tile/src/docs/osds-ovh-location-tile/usage.mdx';
-
-<Meta title="OVH Components/Location Tile [molecule]/Web Component" />
-
-# `<osds-ovh-location-tile>`
-
-## Table of Contents
-> - **[Usage](#usage)**
-> - **[Getting Started](#getting-started)**
-> - **[API Table](#api-table)**
-
-<hr/>
-<Usage />
-<hr/>
-<div id="api-table"><Description>{APITable}</Description></div>
